+++ src/components/MDButton/MDButtonRoot.js
@@ -4,8 +4,8 @@
 * Material Dashboard 3 PRO React - v2.3.0
 =========================================================
 
-* Product Page: https://www.creative-tim.com/product/material-dashboard-pro-react
-* Copyright 2024 Infraspec (https://www.creative-tim.com)
+* Product Page: sytis/product/material-dashboard-pro-react
+* Copyright 2024 Infraspec (sytis)
 
 Coded by www.creative-tim.com